// workflow_fe/src/lib/api/gen/types/RbacRoleResponseDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type RbacRoleResponseDto = {
  /**
   * @type string
   */
  id: string;
  /**
   * @type string
   */
  name: string;
  /**
   * @type string
   */
  slug: string;
  /**
   * @type string
   */
  description: string | null;
  /**
   * @type boolean
   */
  isSystem: boolean;
  /**
   * @type boolean
   */
  isLocked: boolean;
  /**
   * @type array
   */
  permissionSlugs: string[];
  /**
   * @type string
   */
  createdAt: string;
  /**
   * @type string
   */
  updatedAt: string;
};